+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.3.0
+* add support for [uv](https://github.com/astral-sh/uv) and switch the default from pip-tools to uv.
+* improved is-synced check
+* document downloading / offline installation
+* document nix support
+* generally improved documentation
+* add tests
+
 ## 0.2.0
 * bug fixes
 * feature: download dependencies for offline syncing